Clean up transformer models
