+++ xmm_superres_denoise/models/__init__.py
@@ -1,5 +1,2 @@
 from .model import Model
-from .modules.drct import DRCT
 from .modules.generator_rrdb import GeneratorRRDB_DN, GeneratorRRDB_SR
-from .modules.hat import HAT
-from .modules.swinfir import SwinFIR
