+++ xmm_superres_denoise/models/transformer/drct.py
@@ -1,0 +1,328 @@
+import torch
+import torch.nn as nn
+from models.transformer.modules import (
+    PatchEmbed,
+    PatchUnEmbed,
+    SwinTransformerBlock,
+    Upsample,
+)
+from models.transformer.tools import init_weights
+from timm.layers import trunc_normal_
+
+
+class RDG(nn.Module):
+    def __init__(
+        self,
+        dim,
+        input_resolution,
+        num_heads,
+        window_size,
+        mlp_ratio,
+        qkv_bias,
+        qk_scale,
+        drop,
+        attn_drop,
+        drop_path,
+        norm_layer,
+        gc,
+        patch_size,
+        img_size,
+    ):
+        super().__init__()
+
+        self.swin1 = SwinTransformerBlock(
+            dim=dim,
+            input_resolution=input_resolution,
+            num_heads=num_heads,
+            window_size=window_size,
+            shift_size=0,  # For first block
+            mlp_ratio=mlp_ratio,
+            qkv_bias=qkv_bias,
+            qk_scale=qk_scale,
+            drop=drop,
+            attn_drop=attn_drop,
+            drop_path=drop_path[0] if isinstance(drop_path, list) else drop_path,
+            norm_layer=norm_layer,
+        )
+        self.adjust1 = nn.Conv2d(dim, gc, 1)
+
+        self.swin2 = SwinTransformerBlock(
+            dim + gc,
+            input_resolution=input_resolution,
+            num_heads=num_heads - ((dim + gc) % num_heads),
+            window_size=window_size,
+            shift_size=window_size // 2,  # For first block
+            mlp_ratio=mlp_ratio,
+            qkv_bias=qkv_bias,
+            qk_scale=qk_scale,
+            drop=drop,
+            attn_drop=attn_drop,
+            drop_path=drop_path[0] if isinstance(drop_path, list) else drop_path,
+            norm_layer=norm_layer,
+        )
+        self.adjust2 = nn.Conv2d(dim + gc, gc, 1)
+
+        self.swin3 = SwinTransformerBlock(
+            dim + 2 * gc,
+            input_resolution=input_resolution,
+            num_heads=num_heads - ((dim + 2 * gc) % num_heads),
+            window_size=window_size,
+            shift_size=0,  # For first block
+            mlp_ratio=mlp_ratio,
+            qkv_bias=qkv_bias,
+            qk_scale=qk_scale,
+            drop=drop,
+            attn_drop=attn_drop,
+            drop_path=drop_path[0] if isinstance(drop_path, list) else drop_path,
+            norm_layer=norm_layer,
+        )
+        self.adjust3 = nn.Conv2d(dim + gc * 2, gc, 1)
+
+        self.swin4 = SwinTransformerBlock(
+            dim + 3 * gc,
+            input_resolution=input_resolution,
+            num_heads=num_heads - ((dim + 3 * gc) % num_heads),
+            window_size=window_size,
+            shift_size=window_size // 2,  # For first block
+            mlp_ratio=1,
+            qkv_bias=qkv_bias,
+            qk_scale=qk_scale,
+            drop=drop,
+            attn_drop=attn_drop,
+            drop_path=drop_path[0] if isinstance(drop_path, list) else drop_path,
+            norm_layer=norm_layer,
+        )
+        self.adjust4 = nn.Conv2d(dim + gc * 3, gc, 1)
+
+        self.swin5 = SwinTransformerBlock(
+            dim + 4 * gc,
+            input_resolution=input_resolution,
+            num_heads=num_heads - ((dim + 4 * gc) % num_heads),
+            window_size=window_size,
+            shift_size=0,  # For first block
+            mlp_ratio=1,
+            qkv_bias=qkv_bias,
+            qk_scale=qk_scale,
+            drop=drop,
+            attn_drop=attn_drop,
+            drop_path=drop_path[0] if isinstance(drop_path, list) else drop_path,
+            norm_layer=norm_layer,
+        )
+        self.adjust5 = nn.Conv2d(dim + gc * 4, dim, 1)
+
+        self.lrelu = nn.LeakyReLU(negative_slope=0.2, inplace=True)
+
+        self.pe = PatchEmbed(
+            img_size=img_size,
+            patch_size=patch_size,
+            in_chans=0,
+            embed_dim=dim,
+            norm_layer=None,
+        )
+
+        self.pue = PatchUnEmbed(
+            img_size=img_size,
+            patch_size=patch_size,
+            in_chans=0,
+            embed_dim=dim,
+        )
+
+    def forward(self, x, xsize):
+        x1 = self.pe(self.lrelu(self.adjust1(self.pue(self.swin1(x, xsize), xsize))))
+        x2 = self.pe(
+            self.lrelu(
+                self.adjust2(self.pue(self.swin2(torch.cat((x, x1), -1), xsize), xsize))
+            )
+        )
+        x3 = self.pe(
+            self.lrelu(
+                self.adjust3(
+                    self.pue(self.swin3(torch.cat((x, x1, x2), -1), xsize), xsize)
+                )
+            )
+        )
+        x4 = self.pe(
+            self.lrelu(
+                self.adjust4(
+                    self.pue(self.swin4(torch.cat((x, x1, x2, x3), -1), xsize), xsize)
+                )
+            )
+        )
+        x5 = self.pe(
+            self.adjust5(
+                self.pue(self.swin5(torch.cat((x, x1, x2, x3, x4), -1), xsize), xsize)
+            )
+        )
+
+        return x5 * 0.2 + x
+
+
+class DRCT(nn.Module):
+    def __init__(
+        self,
+        img_size=64,
+        patch_size=1,
+        in_chans=3,
+        embed_dim=96,
+        depths=(6, 6, 6, 6),
+        num_heads=(6, 6, 6, 6),
+        window_size=7,
+        overlap_ratio=0.5,
+        mlp_ratio=4.0,
+        qkv_bias=True,
+        qk_scale=None,
+        drop_rate=0.0,
+        attn_drop_rate=0.0,
+        drop_path_rate=0.1,
+        norm_layer=nn.LayerNorm,
+        ape=False,
+        patch_norm=True,
+        use_checkpoint=False,
+        upscale=2,
+        img_range=1.0,
+        upsampler="",
+        resi_connection="1conv",
+        gc=32,
+        **kwargs,
+    ):
+        super(DRCT, self).__init__()
+
+        self.window_size = window_size
+        self.shift_size = window_size // 2
+        self.overlap_ratio = overlap_ratio
+
+        num_in_ch = in_chans
+        num_out_ch = in_chans
+        num_feat = 64
+        self.img_range = img_range
+        if in_chans == 3:
+            rgb_mean = (0.4488, 0.4371, 0.4040)
+            self.mean = torch.Tensor(rgb_mean).view(1, 3, 1, 1)
+        else:
+            self.mean = torch.zeros(1, 1, 1, 1)
+        self.upscale = upscale
+        self.upsampler = upsampler
+
+        # ------------------------- 1, shallow feature extraction ------------------------- #
+        self.conv_first = nn.Conv2d(num_in_ch, embed_dim, 3, 1, 1)
+
+        # ------------------------- 2, deep feature extraction ------------------------- #
+        self.num_layers = len(depths)
+        self.embed_dim = embed_dim
+        self.ape = ape
+        self.patch_norm = patch_norm
+        self.num_features = embed_dim
+        self.mlp_ratio = mlp_ratio
+
+        # split image into non-overlapping patches
+        self.patch_embed = PatchEmbed(
+            img_size=img_size,
+            patch_size=patch_size,
+            in_chans=embed_dim,
+            embed_dim=embed_dim,
+            norm_layer=norm_layer if self.patch_norm else None,
+        )
+        num_patches = self.patch_embed.num_patches
+        patches_resolution = self.patch_embed.patches_resolution
+        self.patches_resolution = patches_resolution
+
+        # merge non-overlapping patches into image
+        self.patch_unembed = PatchUnEmbed(
+            img_size=img_size,
+            patch_size=patch_size,
+            in_chans=embed_dim,
+            embed_dim=embed_dim,
+        )
+
+        # absolute position embedding
+        if self.ape:
+            self.absolute_pos_embed = nn.Parameter(
+                torch.zeros(1, num_patches, embed_dim)
+            )
+            trunc_normal_(self.absolute_pos_embed, std=0.02)
+
+        self.pos_drop = nn.Dropout(p=drop_rate)
+
+        # stochastic depth
+        dpr = [
+            x.item() for x in torch.linspace(0, drop_path_rate, sum(depths))
+        ]  # stochastic depth decay rule
+
+        # build
+        self.layers = nn.ModuleList()
+        for i_layer in range(self.num_layers):
+            layer = RDG(
+                dim=embed_dim,
+                input_resolution=(patches_resolution[0], patches_resolution[1]),
+                num_heads=num_heads[i_layer],
+                window_size=window_size,
+                mlp_ratio=mlp_ratio,
+                qkv_bias=qkv_bias,
+                qk_scale=qk_scale,
+                drop=drop_rate,
+                attn_drop=attn_drop_rate,
+                drop_path=dpr[sum(depths[:i_layer]) : sum(depths[: i_layer + 1])],
+                norm_layer=norm_layer,
+                gc=gc,
+                img_size=img_size,
+                patch_size=patch_size,
+            )
+
+            self.layers.append(layer)
+        self.norm = norm_layer(self.num_features)
+
+        # build the last conv layer in deep feature extraction
+        if resi_connection == "1conv":
+            self.conv_after_body = nn.Conv2d(embed_dim, embed_dim, 3, 1, 1)
+        elif resi_connection == "identity":
+            self.conv_after_body = nn.Identity()
+
+        # ------------------------- 3, high quality image reconstruction ------------------------- #
+        if self.upsampler == "pixelshuffle":
+            # for classical SR
+            self.conv_before_upsample = nn.Sequential(
+                nn.Conv2d(embed_dim, num_feat, 3, 1, 1), nn.LeakyReLU(inplace=True)
+            )
+            self.upsample = Upsample(upscale, num_feat)
+            self.conv_last = nn.Conv2d(num_feat, num_out_ch, 3, 1, 1)
+
+        self.apply(init_weights)
+
+    @torch.jit.ignore
+    def no_weight_decay(self):
+        return {"absolute_pos_embed"}
+
+    @torch.jit.ignore
+    def no_weight_decay_keywords(self):
+        return {"relative_position_bias_table"}
+
+    def forward_features(self, x):
+        x_size = (x.shape[2], x.shape[3])
+
+        x = self.patch_embed(x)
+        if self.ape:
+            x = x + self.absolute_pos_embed
+        x = self.pos_drop(x)
+
+        for layer in self.layers:
+            x = layer(x, x_size)
+
+        x = self.norm(x)  # b seq_len c
+        x = self.patch_unembed(x, x_size)
+
+        return x
+
+    def forward(self, x):
+        self.mean = self.mean.type_as(x)
+        x = (x - self.mean) * self.img_range
+
+        if self.upsampler == "pixelshuffle":
+            # for classical SR
+            x = self.conv_first(x)
+            x = self.conv_after_body(self.forward_features(x)) + x
+            x = self.conv_before_upsample(x)
+            x = self.conv_last(self.upsample(x))
+
+        x = x / self.img_range + self.mean
+
+        return x
